+++ src/stories/components/MainMenu/MainMenu.js
@@ -34,27 +34,37 @@ export const MainMenu = () => {
 
   return (
     <ul className="main-menu">
-      { currentUser ?
-        <li className="main-menu__item">
-          Welcome back,&nbsp;
-          <strong>{profile && profile.username ?
-            profile.username :
-            'Friend'
-          }</strong>!
-        </li> :
-        <li className="main-menu__item">
-          Hello there,&nbsp;
-          <strong>Stranger</strong>!
-        </li>
+      {
+        currentUser ?
+          <li className="main-menu__item">
+            Welcome back,&nbsp;
+            <strong>{profile && profile.username ?
+              profile.username :
+              'Friend'
+            }</strong>!
+          </li> :
+          <li className="main-menu__item">
+            Hello there,&nbsp;
+            <strong>Stranger</strong>!
+          </li>
       }
       <li className="main-menu__item">
         <Link to="/search"
           className="main-menu__link main-menu__link--search">Search</Link>
       </li>
+      {
+        currentUser && profile && profile.userType === 'Chef' &&
+        <li className="main-menu__item">
+
+          <Link to="/add_recipe"
+            className="main-menu__link main-menu__link--add">Add</Link>
+        </li>
+      }
       <li className="main-menu__item">
-        { currentUser ?
-          <Button url="/account/information" text="My Account" /> :
-          <Button url="/login" text="Login" />
+        {
+          currentUser ?
+            <Button url="/account/information" text="My Account"/> :
+            <Button url="/login" text="Login"/>
         }
       </li>
     </ul>
