+++ src/stories/components/MainMenu/styles.scss
@@ -36,4 +36,18 @@
       text-decoration: underline;
     }
   }
+
+  &.main-menu__link--add {
+    @include svg-background(icon-add-post);
+
+    background-position: left center;
+    background-repeat: no-repeat;
+    font-weight: 400;
+    padding: $spacing-1 0  $spacing-1 $spacing-3;
+
+    &:hover,
+    &:focus {
+      text-decoration: underline;
+    }
+  }
 }
