--- src/stories/components/NewRecipe/styles.scss
@@ -1,4 +1,0 @@
-@import '../../../scss/variables.scss';
-
-// @file
-// Review
